feat(detection): LSTM sequence model, Isolation Forest, ensemble & pipeline
Tasks 9-12:
- LstmSequenceDetector (PyTorch DeepLog): next-event prediction, top-k surprise,
  perplexity, predicted-vs-actual + suspicious subsequence; safe weights-only load
- IsolationForestDetector: decision-function sigmoid calibration (boundary at 0.5),
  occlusion-based per-feature attribution; artifact save/load
- WeightedEnsemble (normalized weights, threshold, severity banding) + Explainer
- DetectAnomalies use case: score -> evidence -> alert -> publish AlertCreated
17 detection tests (unit + ML integration); 92 total green, mypy --strict clean.

+++ backend/src/nexguard/domain/value_objects.py
@@ -143,7 +143,7 @@ class CountVector:
 
     @classmethod
     def from_counts(
-        cls, counts: Mapping[EventId, int], vocab: Sequence[EventId]
+        cls, counts: Mapping[EventId, float], vocab: Sequence[EventId]
     ) -> CountVector:
         ordered_vocab = tuple(vocab)
         values = tuple(float(counts.get(event_id, 0)) for event_id in ordered_vocab)
